+++ src/ksp/ksp.c
@@ -197,9 +197,9 @@ _pgr_ksp(PG_FUNCTION_ARGS) {
         }
         int64_t seq = funcctx->call_cntr == 0?  1 : path[funcctx->call_cntr - 1].end_id;
 
-        values[0] = Int32GetDatum(funcctx->call_cntr + 1);
-        values[1] = Int32GetDatum(path_id);
-        values[2] = Int32GetDatum(seq);
+        values[0] = Int32GetDatum((int32_t)funcctx->call_cntr + 1);
+        values[1] = Int32GetDatum((int32_t)path_id);
+        values[2] = Int32GetDatum((int32_t)seq);
         if (PG_NARGS() != 6) {
             values[3] = Int64GetDatum(path[funcctx->call_cntr].start_id);
             values[4] = Int64GetDatum(path[funcctx->call_cntr].end_id);
